move muteIndicatorViewController form callViewController to VideoGridViewController
